--- shopperpay/demo/gs_query_demo.php
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<title>GS Query Demo</title>
-</head>
-<body style="text-align:center">
-    <p>GS Order Search Demo </p>
-    <form action="../order_gs_query.php" method="post">
-        GSOrdId: <input type="text" name="GSOrdId" value="1471681439851441">
-        MerOrdId: <input type="text" name="MerOrdId" value="2016082016211324">
-    <input type="submit"  value="order search">
-    </form>
-</body>
-</html>
-
