--- shopperpay/demo/return_url_demo.php
@@ -1,2 +1,0 @@
-<?php
-var_dump($_POST);
