+++ shopperpay/pay/demo/seller_api_demo.php
@@ -1,0 +1,7 @@
+<?php
+/*
+ *  商户提供的
+ */
+echo json_encode(array('isSuccess'=>'1', 'errno'=>'', 'errMessage'=>''));
+
+
